Web search plugins split
the plugins websearch and moreserach were separated in separate
plugins, to make it easier to adapt each plugin for each type of
search, simplify the search for problems or conflicts, it is cleaner
and gives each user the ability to chose what search it prefferes.

also there will be only one imdb search, after comparing the files
there were no need for 2

moresearch.bundle and webserach.bundle were deleted.