Change API, so that more currencies are available.
This changes the API to bitcoinaverage to support a wider range of
currencies.

The terms of use of this API are at https://bitcoinaverage.com/api,
which state the API is free to use.

Since more currencies are available, we add a preference to select
which to use (list manually fetched from api.bitcoinaverage.com/ticker)
and also to automatically choose using the same method as the CNY
currency currently uses.

Currency symbols were acquired from xe.com.

Commented out and dead code was also removed in the process.