Change API from EarthTool to Google Time Zone
The EarthTool API does not seem to always know daylight savings for a
location. This causes the incorrect time and offset to be reported.

This changes to using the [Google Time Zone
API](https://developers.google.com/maps/documentation/timezone), which
knows daylight savings and provides it separately for us to take into
account.

The free API allows 2,500 requests per day, and 5 per second, which
should be more than enough.