Fix localisation by searching first, then select correct country.
We make an initial call to the /find endpoint to get all results, then get the ID of the first one with the correct country code.
If this fails, then we just get the weather of the first result.
This should ensure priority of cities in the same country as the user, and also allows the user to override this by putting location,country instead where country is an ISO code.
It does add some overhead however, as we have now doubled the amount of API calls we need to make.