Change lock command to not use fast user switching.
This commit changes to "pressing" the lock screen button on the Keychain menu by importing the menu as a bundle and calling the necessary function.
See http://apple.stackexchange.com/a/123738 for details.

We add another argument to run() to select this. A bit messy, but results in the smallest net change. **kwargs would be better, but don't seem to work.